Significant changes to aid UDP connections.  * Added Period and Delay parameters to connections.    - Delay is used to avoid simultaneous TCP connections.    - Period is used to regulate UDP connections.  * added Delay code to pqissl.  * added Period code to tcponudp / pqissludp.  * modified TTL modification code.  * increased SynPktRetransmit value.  * fixed retrans() timeout (one reason code wasn't working before!)  * fixed tou_close() SEGV bug.  * modified pqissludp tou_socket creation. (non permanent now).  * Modified format of peerConnectRequest() CB to make it more useful and rewrote function.  * Enabled pqissludp NetInterface.  * using Id comparision to determine Active/Passive UDP connection state.  * added #def to disable TCP connections. (for testing)  * enabled UDP connections from retryConnect() function.  * corrected EXT check in retryConnect() function.  * + lots of debug output and other stuff.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@358 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/tcponudp/tcppacket.h
@@ -60,7 +60,7 @@ class TcpPacket
 
 	/* other variables */
 	double  ts; /* transmit time */ 
-	uint8   retrans; /* retransmit counter */
+	uint16  retrans; /* retransmit counter */
 
 	TcpPacket(uint8 *ptr, int size);
 	TcpPacket(); /* likely control packet */
